# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "T&C API",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/clinics/{clinic}": {
            "get": {
                "tags": [
                    "Clinics"
                ],
                "summary": "Get a Specific Clinic",
                "description": "Displays information about a specific clinic.",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "clinic",
                        "in": "path",
                        "description": "ID of the clinic",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clinic fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Clinic fetched successfully!"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ClinicResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found - Clinic not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Clinic not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clinics"
                ],
                "summary": "Delete a Clinic",
                "description": "Removes the specified clinic.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "clinic",
                        "in": "path",
                        "description": "ID of the clinic",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clinic deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Clinic deleted successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found - Clinic not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Clinic not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Clinics"
                ],
                "summary": "Update a Clinic",
                "description": "Updates the specified clinic.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "clinic",
                        "in": "path",
                        "description": "ID of the clinic",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated clinic details",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Clinic Name"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Updated Clinic Address"
                                    },
                                    "postcode": {
                                        "type": "string",
                                        "example": "54321"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Updated Clinic City"
                                    },
                                    "latitude": {
                                        "type": "string",
                                        "example": "Updated Clinic Latitude"
                                    },
                                    "longitude": {
                                        "type": "string",
                                        "example": "Updated Clinic Longitude"
                                    },
                                    "country_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Clinic updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Clinic updated successfully!"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ClinicResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found - Clinic not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Clinic not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity - Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/clinic-users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Store Clinic User",
                "description": "Stores a new clinic user.",
                "operationId": "storeClinicUser",
                "requestBody": {
                    "description": "Clinic user details",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "role",
                                    "clinic_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "clinic_owner"
                                    },
                                    "clinic_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Clinic user created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "clinic_owner"
                                        },
                                        "clinic_id": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity - Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClinicResource": {
                "title": "Clinic Resource",
                "description": "Represents a clinic in the system.",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the clinic.",
                        "type": "string"
                    },
                    "name": {
                        "description": "The name of the clinic.",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the clinic.",
                        "type": "string"
                    },
                    "address": {
                        "description": "The address of the clinic.",
                        "type": "string"
                    },
                    "postcode": {
                        "description": "The postcode of the clinic.",
                        "type": "string"
                    },
                    "city": {
                        "description": "The city where the clinic is located.",
                        "type": "string"
                    },
                    "country_name": {
                        "description": "The name of the country where the clinic is located.",
                        "type": "string"
                    },
                    "phone": {
                        "description": "The phone number of the clinic.",
                        "type": "string"
                    },
                    "mobile": {
                        "description": "The mobile phone number of the clinic.",
                        "type": "string"
                    },
                    "email": {
                        "description": "The email address of the clinic.",
                        "type": "string",
                        "format": "email"
                    },
                    "web": {
                        "description": "The website of the clinic.",
                        "type": "string"
                    },
                    "contact_person": {
                        "description": "The contact person of the clinic.",
                        "type": "string"
                    },
                    "contact_phone": {
                        "description": "The contact phone number of the clinic.",
                        "type": "string"
                    },
                    "contact_email": {
                        "description": "The contact email address of the clinic.",
                        "type": "string",
                        "format": "email"
                    },
                    "latitude": {
                        "description": "The latitude of the clinic location.",
                        "type": "string"
                    },
                    "longitude": {
                        "description": "The longitude of the clinic location.",
                        "type": "string"
                    },
                    "created_by": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "updated_by": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "logo_image": {
                        "description": "The path to the logo image of the clinic.",
                        "type": "string"
                    },
                    "categories": {
                        "description": "The categories associated with the clinic.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CategoryResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}